Deleting a User's task

+++ src/tasks/tasks.service.ts
@@ -53,8 +53,11 @@ export class TasksService {
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(
+    id: number,
+    user: User
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (!result.affected) {
       throw new NotFoundException(`Task ${id} not found`);
